:tropical_drink: No profile, os eventos só se apagam com token de sessão correcto

+++ Source Code/PHP/event_deletion.php
@@ -2,7 +2,7 @@
 
 if (isset($_POST['Delete_button'])) {
 	$event_to_delete = $_POST['Delete_button'];
-	if (existsEventByID($event_to_delete))	
+	if (existsEventByID($event_to_delete) && $_SESSION['csrf'] == $_POST['csrf'])	
 	{
 		$event_name =  deleteEvent($event_to_delete);
 	}
